Add files via upload
Class Coordinate: It helps to distinguish each lattice.
Class MyButton : This class inherit the class Button. It helps to define a type of button which permit add color as one of its attributes.
Class Question: This class permit to set a question when user click on the button "change a question".
Class TestGame: This class contains the main method.